refactor(watch): carve events.rs out of mod.rs
src/cli/watch/events.rs (~440 lines):
- max_pending_files (env-var read, hot-cached)
- collect_events: per-event filter + dedupe into WatchState pending sets
- process_file_changes: drains pending_files into reindex + HNSW maintenance
- process_note_changes: drains pending_notes into reindex_notes

Uses 'use super::*;' since the dispatch loop reaches across the rebuild,
gc, reindex, and shared-types surfaces. max_pending_files re-import in
mod.rs is #[cfg(test)] gated — only the test bench reads it directly.
